adjustments double to BigDecimal

+++ producer/src/main/java/com/order/service/api/producer/model/Product.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @AllArgsConstructor
@@ -19,6 +21,7 @@ public class Product {
     private String name;
 
     @Column(nullable = false)
-    private double unitPrice;
+    private BigDecimal unitPrice;
+
 
 }
